Move cli tests to ui_tests/cli
