+++ tests/ui_tests/cli/analysis/snapshots/test_es_update/test_update_only_using_subset_observations/0/update_log
