+++ tests/ui_tests/cli/analysis/test_adaptive_localization.py
@@ -9,7 +9,7 @@ from ert.mode_definitions import ENSEMBLE_SMOOTHER_MODE
 from ert.storage import open_storage
 
 random_seed_line = "RANDOM_SEED 1234\n\n"
-from tests.integration_tests.run_cli import run_cli
+from tests.ui_tests.cli.run_cli import run_cli
 
 
 def run_cli_ES_with_case(poly_config):
