+++ tests/ui_tests/cli/run_cli.py
