+++ tests/ui_tests/cli/snapshots/test_cli/test_es_mda/using_scheduler/es_mda_integration_snapshot
