+++ tests/ui_tests/cli/test_cli.py
