+++ tests/ui_tests/cli/test_observation_times.py
